xnu-4903.221.2
Imported from https://opensource.apple.com/tarballs/xnu/xnu-4903.221.2.tar.gz

+++ libsyscall/wrappers/init_cpu_capabilities.c
@@ -35,7 +35,7 @@ int _cpu_capabilities = 0;
 void
 _init_cpu_capabilities( void )
 {
-	_cpu_capabilities = _get_cpu_capabilities();
+	_cpu_capabilities = (int)_get_cpu_capabilities();
 }
 
 #elif defined(__arm__) || defined(__arm64__)
